refactor: move `get_ext_stac_collection` to `eodag.utils`
