+++ tests/units/test_stac_utils.py
@@ -253,8 +253,7 @@ class TestStacUtils(unittest.TestCase):
         self.assertEqual(dtstart, start)
         self.assertEqual(dtend, end)
 
-    @mock.patch("eodag.rest.stac.get_ext_stac_collection", autospec=True)
-    def test_fetch_external_stac_collections(self, mock_get_ext_stac_collection):
+    def test_fetch_external_stac_collections(self):
         """Load external STAC collections"""
         external_json = """{
             "new_field":"New Value",
@@ -275,13 +274,11 @@ class TestStacUtils(unittest.TestCase):
                 ]
             }
         }"""
-        mock_get_ext_stac_collection.return_value = json.loads(external_json)
         product_type_conf = self.rest_core.eodag_api.product_types_config["S2_MSI_L1C"]
         ext_stac_collection_path = "/path/to/external/stac/collections/S2_MSI_L1C.json"
         product_type_conf["stacCollection"] = ext_stac_collection_path
-        # Update EODataAccessGateway.product_types_config
-        # Using a context manager cause `self.rest.core.get_stac_collection_by_id needs`
-        # to call the not-patched version of `eodag.api.core.get_ext_stac_collection`
+
+        # Check if `eodag.utils.get_ext_stac_collection()` is called with right arg
         with mock.patch(
             "eodag.api.core.get_ext_stac_collection", autospec=True
         ) as mock_core_get_ext_stac_collection:
@@ -290,25 +287,32 @@ class TestStacUtils(unittest.TestCase):
             mock_core_get_ext_stac_collection.assert_called_once_with(
                 ext_stac_collection_path
             )
-        # Init StacCollection.ext_stac_collections
-        StacCollection.fetch_external_stac_collections(self.rest_core.eodag_api)
-        r = self.rest_core.get_stac_collection_by_id(
-            url="", root="", collection_id="S2_MSI_L1C"
-        )
-        mock_get_ext_stac_collection.assert_called_with(ext_stac_collection_path)
-        # Fields not supported by EODAG
-        self.assertIn("new_field", r)
-        # Merge lists
-        self.assertEqual(
-            r["summaries"]["platform"][0], "S2A,S2B,SENTINEL-2A,SENTINEL-2B"
-        )
-        # Don't override existing keys
-        self.assertEqual(r["title"], "SENTINEL2 Level-1C")
-        # Restore previous state
-        del self.rest_core.eodag_api.product_types_config["S2_MSI_L1C"][
-            "stacCollection"
-        ]
-        StacCollection.ext_stac_collections.clear()
+
+        # Check if the returned STAC collection contains updated data
+        with mock.patch(
+            "eodag.rest.stac.get_ext_stac_collection", autospec=True
+        ) as mock_stac_get_ext_stac_collection:
+            mock_stac_get_ext_stac_collection.return_value = json.loads(external_json)
+            StacCollection.fetch_external_stac_collections(self.rest_core.eodag_api)
+            r = self.rest_core.get_stac_collection_by_id(
+                url="", root="", collection_id="S2_MSI_L1C"
+            )
+            mock_stac_get_ext_stac_collection.assert_called_with(
+                ext_stac_collection_path
+            )
+            # Fields not supported by EODAG
+            self.assertIn("new_field", r)
+            # Merge lists
+            self.assertEqual(
+                r["summaries"]["platform"][0], "S2A,S2B,SENTINEL-2A,SENTINEL-2B"
+            )
+            # Don't override existing keys
+            self.assertEqual(r["title"], "SENTINEL2 Level-1C")
+            # Restore previous state
+            del self.rest_core.eodag_api.product_types_config["S2_MSI_L1C"][
+                "stacCollection"
+            ]
+            StacCollection.ext_stac_collections.clear()
 
 
 class TestEodagCql2jsonEvaluator(unittest.TestCase):
